Reorganized test classes
