+++ src/test/java/ConversionsTest.java
@@ -7,13 +7,14 @@ import formulaa.foltl.semantics.FoLtlAssignment;
 import net.sf.tweety.logics.commons.syntax.Constant;
 import net.sf.tweety.logics.fol.syntax.FolFormula;
 import net.sf.tweety.logics.fol.syntax.FolSignature;
-import net.sf.tweety.logics.pl.sat.Sat4jSolver;
 import net.sf.tweety.logics.pl.syntax.PropositionalFormula;
-import org.junit.Assert;
-import org.junit.Test;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.LinkedHashSet;
+
+//import static org.junit.Assert.*;
+import org.junit.Assert;
+import org.junit.Test;
 
 /**
  * Created by Simone Calciolari on 31/08/15.
@@ -141,339 +142,6 @@ public class ConversionsTest {
 		computed.assignSort(new FoLtlVariable("x"), sortAB);
 		expected = parseLTLfFormula("(pbc || pbb || pba) || (pac || pab || paa)");
 		assertEquals("", expected, computed.propositionalize(domain, assignment));
-	}
-
-	@Test
-	public void testQuantifierExpansion(){
-		System.out.println("*** QUANTIFIER EXPANSION TEST ***\n");
-
-		//FoLtlAssignment assignment = new FoLtlAssignment();
-		LinkedHashSet<FoLtlConstant> domain = new LinkedHashSet<>();
-
-		//Atom conversions
-		FoLtlLocalFormula computed = (FoLtlLocalFormula) parseFoLtlFormula("P(a)");
-		FoLtlFormula expected = parseFoLtlFormula("P(a)");
-		assertEquals("", expected, computed.quantifierExpansion(domain));
-
-		computed = (FoLtlLocalFormula) parseFoLtlFormula("TRUE");
-		expected = parseFoLtlFormula("TRUE");
-		assertEquals("", expected, computed.quantifierExpansion(domain));
-
-		computed = (FoLtlLocalFormula) parseFoLtlFormula("FALSE");
-		expected = parseFoLtlFormula("FALSE");
-		assertEquals("", expected, computed.quantifierExpansion(domain));
-
-		computed = (FoLtlLocalFormula) parseFoLtlFormula("a = b");
-		expected = parseFoLtlFormula("a = b");
-		assertEquals("", expected, computed.quantifierExpansion(domain));
-
-		computed = (FoLtlLocalFormula) parseFoLtlFormula("a = a");
-		expected = parseFoLtlFormula("a = a");
-		assertEquals("", expected, computed.quantifierExpansion(domain));
-
-		//Simple boolean conversions
-		computed = (FoLtlLocalFormula) parseFoLtlFormula("! P(b)");
-		expected = parseFoLtlFormula("! P(b)");
-		assertEquals("", expected, computed.quantifierExpansion(domain));
-
-		computed = (FoLtlLocalFormula) parseFoLtlFormula("P(a) && P(b)");
-		expected = parseFoLtlFormula("P(a) && P(b)");
-		assertEquals("", expected, computed.quantifierExpansion(domain));
-
-		computed = (FoLtlLocalFormula) parseFoLtlFormula("P(a) | P(b)");
-		expected = parseFoLtlFormula("P(a) || P(b)");
-		assertEquals("", expected, computed.quantifierExpansion(domain));
-
-		computed = (FoLtlLocalFormula) parseFoLtlFormula("P(a) -> P(b)");
-		expected = parseFoLtlFormula("P(a) -> P(b)");
-		assertEquals("", expected, computed.quantifierExpansion(domain));
-
-		computed = (FoLtlLocalFormula) parseFoLtlFormula("P(a) <-> P(b)");
-		expected = parseFoLtlFormula("P(a) <-> P(b)");
-		assertEquals("", expected, computed.quantifierExpansion(domain));
-
-		//More intricate boolean conversions
-		computed = (FoLtlLocalFormula) parseFoLtlFormula("P(p) || Q(a) && ! V(r) -> J(s)");
-		expected = parseFoLtlFormula("P(p) || Q(a) && ! V(r) -> J(s)");
-		assertEquals("", expected, computed.quantifierExpansion(domain));
-
-		computed = (FoLtlLocalFormula) parseFoLtlFormula("P(a) -> P(a) -> P(a) -> P(a)");
-		expected = parseFoLtlFormula("P(a) -> (P(a) -> (P(a) -> P(a)))");
-		assertEquals("", expected, computed.quantifierExpansion(domain));
-
-		computed = (FoLtlLocalFormula) parseFoLtlFormula("!(!(!P(a)))");
-		expected = parseFoLtlFormula("!(!(!P(a)))");
-		assertEquals("", expected, computed.quantifierExpansion(domain));
-
-		computed = (FoLtlLocalFormula) parseFoLtlFormula("P(a) && (P(c) || P(b) || P(d)) -> P(s) && P(rst)");
-		expected = parseFoLtlFormula("P(a) && (P(c) || P(b) || P(d)) -> P(s) && P(rst)");
-		assertEquals("", expected, computed.quantifierExpansion(domain));
-
-		//Testing quantifiers and substitutions
-		domain.add(new FoLtlConstant("a"));
-		domain.add(new FoLtlConstant("b"));
-
-		String input = "Exists ?x P(?x)";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		expected = parseFoLtlFormula("P(b) || P(a)");
-		assertEquals(input, expected, computed.quantifierExpansion(domain));
-
-		input = "Forall ?x P(?x)";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		expected = parseFoLtlFormula("P(b) && P(a)");
-		assertEquals(input, expected, computed.quantifierExpansion(domain));
-
-		input = "Exists ?x P(?x) && P(d)";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		expected = parseFoLtlFormula("(P(b) && P(d)) || (P(a) && P(d))");
-		assertEquals(input, expected, computed.quantifierExpansion(domain));
-
-		input = "Forall ?x P(?x) -> P(?z)";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		expected = parseFoLtlFormula("(P(b) -> P(?z)) && (P(a) -> P(?z))");
-		assertEquals(input, expected, computed.quantifierExpansion(domain));
-
-		input = "Exists ?x (Exists ?y (P(?x) && P(?y)))";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		expected = parseFoLtlFormula("(P(b) & P(b) || (P(b) & P(a))) || ((P(a) & P(b)) | (P(a) & P(a)))");
-		assertEquals(input, expected, computed.quantifierExpansion(domain));
-
-
-		//Testing expansion with sorts
-		FoLtlSort sortAB = new FoLtlSort("AB");
-		sortAB.add(new FoLtlConstant("a"));
-		sortAB.add(new FoLtlConstant("b"));
-
-		FoLtlSort sortC = new FoLtlSort("C");
-		sortC.add(new FoLtlConstant("c"));
-
-		domain.add(new FoLtlConstant("c"));
-
-		System.out.println("\nSORT EXPANSION TEST\n");
-
-		input = "Exists ?x P(?x)";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		computed.assignSort(new FoLtlVariable("x"), sortAB);
-		expected = parseFoLtlFormula("P(b) || P(a)");
-		assertEquals(input, expected, computed.quantifierExpansion(domain));
-
-		input = "Exists ?x P(?x)";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		computed.assignSort(new FoLtlVariable("x"), sortC);
-		expected = parseFoLtlFormula("P(c)");
-		assertEquals(input, expected, computed.quantifierExpansion(domain));
-
-		input = "Exists ?x (Exists ?y P(?x, ?y))";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		computed.assignSort(new FoLtlVariable("x"), sortAB);
-		expected = parseFoLtlFormula("(P(b, c) || P(b, b) || P(b, a)) || (P(a, c) || P(a, b) || P(a, a))");
-		assertEquals(input, expected, computed.quantifierExpansion(domain));
-	}
-
-	@Test
-	public void testStisfiability(){
-		System.out.println("*** SATISFIABILITY TEST *** \n");
-
-		LinkedHashSet<FoLtlConstant> domain = new LinkedHashSet<>();
-		domain.add(new FoLtlConstant("a"));
-		domain.add(new FoLtlConstant("b"));
-
-		System.out.println("Domain: " + domain.toString() + "\n");
-
-		FoLtlAssignment assignment = new FoLtlAssignment();
-
-		//Simple satisfiable formulas
-		String input = "P(a)";
-		FoLtlLocalFormula computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		assertTrue(computed + " is satisfiable", computed.isSatisfiable(domain, assignment));
-
-		input = "TRUE";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		assertTrue(computed + " is satisfiable", computed.isSatisfiable(domain, assignment));
-
-		input = "a = a";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		assertTrue(computed + " is satisfiable", computed.isSatisfiable(domain, assignment));
-
-		input = "! P(b)";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		assertTrue(computed + " is satisfiable", computed.isSatisfiable(domain, assignment));
-
-		input = "P(a) && P(b)";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		assertTrue(computed + " is satisfiable", computed.isSatisfiable(domain, assignment));
-
-		input = "P(a) && ! P(b)";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		assertTrue(computed + " is satisfiable", computed.isSatisfiable(domain, assignment));
-
-		input = "P(a) | P(b)";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		assertTrue(computed + " is satisfiable", computed.isSatisfiable(domain, assignment));
-
-		input = "P(a) -> P(b)";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		assertTrue(computed + " is satisfiable", computed.isSatisfiable(domain, assignment));
-
-		input = "P(a) <-> P(b)";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		assertTrue(computed + " is satisfiable", computed.isSatisfiable(domain, assignment));
-
-		input = "((P(a) -> Q(a)) -> P(a)) -> P(a)";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		assertTrue(computed + " is satisfiable", computed.isSatisfiable(domain, assignment));
-
-		input = "! (P(a) | Q(a) -> P(a) & Q(a))";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		assertTrue(computed + " is satisfiable", computed.isSatisfiable(domain, assignment));
-
-		input = "(P(a) <-> Q(a)) -> (!Q(a) <-> P(a))";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		assertTrue(computed + " is satisfiable", computed.isSatisfiable(domain, assignment));
-
-		input = "(C(a) -> A(a)) & (!C(a) -> B(a)) & (!(A(a) -> !B(a)))";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		assertTrue(computed + " is satisfiable", computed.isSatisfiable(domain, assignment));
-
-		input = "(P(a) || K(a)) & (!K(a) | !V(a)) & (Q(a) | !V(a)) & (!Q(a) | S(a)) " +
-				"& (!S(a) | !K(a) | M(a)) & (!M(a) | K(a) | !S(a))";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		assertTrue(computed + " is satisfiable", computed.isSatisfiable(domain, assignment));
-
-		input = "(A(a) | B(a) | D(a)) & (!A(a) | B(a) | !C(a)) " +
-				"& (!A(a) | C(a) | D(a)) & (!A(a) | !B(a) | C(a))";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		assertTrue(computed + " is satisfiable", computed.isSatisfiable(domain, assignment));
-
-		input = "!((A(a) & B(a) -> C(a)) -> (A(a) -> C(a)))";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		assertTrue(computed + " is satisfiable", computed.isSatisfiable(domain, assignment));
-
-
-		//Simple unsatisfiable formulas
-		input = "false";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		assertFalse(computed + " is unsatisfiable", computed.isSatisfiable(domain, assignment));
-
-		input = "a = b";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		assertFalse(computed + " is unsatisfiable", computed.isSatisfiable(domain, assignment));
-
-		input = "P(a) & !P(a)";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		assertFalse(computed + " is unsatisfiable", computed.isSatisfiable(domain, assignment));
-
-		input = "!P(a) <-> P(a)";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		assertFalse(computed + " is unsatisfiable", computed.isSatisfiable(domain, assignment));
-
-		input = "! (((P(a) -> Q(b)) -> P(a)) -> P(a))";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		assertFalse(computed + " is unsatisfiable", computed.isSatisfiable(domain, assignment));
-
-		input = "! (((A(a) & B(a)) -> C(a)) -> ((A(a) -> C(a)) | (B(a) -> C(a))))";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		assertFalse(computed + " is unsatisfiable", computed.isSatisfiable(domain, assignment));
-
-		input = "! (((A(a) -> B(a)) -> A(a)) -> A(a))";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		assertFalse(computed + " is unsatisfiable", computed.isSatisfiable(domain, assignment));
-
-		input = "! ((A(a) -> B(a)) -> (!A(a) | B(a)))";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		assertFalse(computed + " is unsatisfiable", computed.isSatisfiable(domain, assignment));
-
-		input = "! ((A(a) -> B(a)) -> (!A(a) | B(a)))";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		assertFalse(computed + " is unsatisfiable", computed.isSatisfiable(domain, assignment));
-
-		input = "!(!(A(a) -> B(a)) -> (A(a) & !B(a)))";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		assertFalse(computed + " is unsatisfiable", computed.isSatisfiable(domain, assignment));
-
-
-		//Satisfiable formulas with quantifiers
-		input = "Exists ?x P(?x)";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		assertTrue(computed + " is satisfiable", computed.isSatisfiable(domain, assignment));
-
-		input = "Forall ?x P(?x)";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		assertTrue(computed + " is satisfiable", computed.isSatisfiable(domain, assignment));
-
-		input = "Exists ?x (Exists ?y (P(?x) && P(?y)))";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		assertTrue(computed + " is satisfiable", computed.isSatisfiable(domain, assignment));
-
-		//D = {1, 2}
-		input = "!((Forall ?x (Exists ?y (P(?x, ?y)))) -> (Forall ?y (Exists ?x (P(?x, ?y)))))";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		assertTrue(computed + " is satisfiable", computed.isSatisfiable(domain, assignment));
-
-		input = "((Forall ?x (Forall ?y (P(?x) & P(?y) -> Q(?x, ?y))))" +
-				" -> (Forall ?x (P(?x) -> (Exists ?y (Q(?x, ?y))))))";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		assertTrue(computed + " is satisfiable", computed.isSatisfiable(domain, assignment));
-
-		input = "(P(a, b) <-> (Exists ?x (Exists ?y (P(?x, ?y)))))";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		assertTrue(computed + " is satisfiable", computed.isSatisfiable(domain, assignment));
-
-		input = "((Forall ?x (Forall ?y (Forall ?z (P(?x, ?y) & P(?x, ?z) -> P(?y, ?z))))) " +
-				"-> (Forall ?x ((Exists ?w (P(?w, ?x))) -> P(?x, ?x))))";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		assertTrue(computed + " is satisfiable", computed.isSatisfiable(domain, assignment));
-
-
-		//Unsatisfiable formulas with quantifiers
-		input = "(Forall ?x P(?x)) && !P(a)";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		assertFalse(computed + " is unsatisfiable", computed.isSatisfiable(domain, assignment));
-
-		input = "!((Forall ?x (Forall ?y (P(?x) & P(?y) -> Q(?x, ?y))))" +
-				" -> (Forall ?x (P(?x) -> (Exists ?y (Q(?x, ?y))))))";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		assertFalse(computed + " is unsatisfiable", computed.isSatisfiable(domain, assignment));
-
-		input = "!((Forall ?x (Forall ?y (Forall ?z (P(?x, ?y) & P(?x, ?z) -> P(?y, ?z))))) " +
-				"-> (Forall ?x ((Exists ?w (P(?w, ?x))) -> P(?x, ?x))))";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		assertFalse(computed + " is unsatisfiable", computed.isSatisfiable(domain, assignment));
-
-
-
-
-
-
-		/*
-		//Testing expansion with sorts
-		FoLtlSort sortAB = new FoLtlSort("AB");
-		sortAB.add(new FoLtlConstant("a"));
-		sortAB.add(new FoLtlConstant("b"));
-
-		FoLtlSort sortC = new FoLtlSort("C");
-		sortC.add(new FoLtlConstant("c"));
-
-		System.out.println("\nSORT EXPANSION TEST\n");
-
-		input = "Exists ?x P(?x)";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		computed.assignSort(new FoLtlVariable("x"), sortAB);
-		expected = parseFoLtlFormula("P(b) || P(a)");
-		assertEquals(input, expected, computed.quantifierExpansion(domain));
-
-		input = "Exists ?x P(?x)";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		computed.assignSort(new FoLtlVariable("x"), sortC);
-		expected = parseFoLtlFormula("P(c)");
-		assertEquals(input, expected, computed.quantifierExpansion(domain));
-
-		input = "Exists ?x (Exists ?y P(?x, ?y))";
-		computed = (FoLtlLocalFormula) parseFoLtlFormula(input);
-		computed.assignSort(new FoLtlVariable("x"), sortAB);
-		expected = parseFoLtlFormula("(P(b, c) || P(b, b) || P(b, a)) || (P(a, c) || P(a, b) || P(a, a))");
-		assertEquals(input, expected, computed.quantifierExpansion(domain));
-		*/
 	}
 
 	@Test
